Completed

+++ src/pages/Repositories/index.js
@@ -1,25 +1,33 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styled';
+import { useNavigate } from 'react-router-dom';
 
 export default function Repositories() {
+  const navigate = useNavigate();
   const [ repositories, setRepositories ] = useState([]);
   useEffect(() => {
     let repositoriesName = localStorage.getItem('repositoriesName');
-    repositoriesName = JSON.parse(repositoriesName);
-    setRepositories(repositoriesName);
-    localStorage.clear();
+    
+    if(repositoriesName != null){
+      repositoriesName = JSON.parse(repositoriesName);
+      setRepositories(repositoriesName);
+      localStorage.clear();
+    } else {
+      navigate('/');
+    }
   }, [])
   
   return(
     <S.Container>
       <S.Title>Repositories</S.Title>
       <S.List>
-        { repositories.map(repository => {
+        { repositories.map((repository, index) => {
           return (
-            <S.listItem>Repositorio: { repository }</S.listItem>
+            <S.listItem key={ index }>Repositorio: { repository }</S.listItem>
           )
         }) };
       </S.List>
+      <S.LinkHome to="/">Voltar</S.LinkHome>
     </S.Container>
   )
 }
